# src/apps/mooc/backend/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ErrorMessage": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserPutBody": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"email",
					"id"
				],
				"type": "object"
			},
			"UserResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserSession": {
				"properties": {
					"id": {
						"type": "string"
					},
					"user": {
						"properties": {
							"email": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"email",
							"id"
						],
						"type": "object"
					},
					"expiration": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"user",
					"expiration"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetSessionResponse": {
				"properties": {
					"session": {
						"$ref": "#/components/schemas/IUserSession"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"session",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSessionPostRequest": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"email"
				],
				"type": "object"
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"name": "jwt",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "socialnetwork-backend",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/v1/status": {
			"get": {
				"operationId": "Run",
				"responses": {
					"200": {
						"description": "Status ok"
					}
				},
				"tags": [
					"Status"
				],
				"security": [],
				"parameters": []
			}
		},
		"/v1/users": {
			"put": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created"
					},
					"404": {
						"description": "Invalid user parameters",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserPutBody"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "SearchUsers",
				"responses": {
					"200": {
						"description": "Users list",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "filters",
						"required": false,
						"schema": {
							"type": "array",
							"items": {}
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/users/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Users list",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/users/session": {
			"post": {
				"operationId": "CreateUserSession",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetSessionResponse"
								}
							}
						}
					},
					"404": {
						"description": "Invalid user parameters",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserSessionPostRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}